refactor: 중형 테스트 medium으로 분리
