--- src/test/java/backend/sumnail/domain/auth/controller/AuthControllerTest.java
@@ -1,7 +1,0 @@
-package backend.sumnail.domain.auth.controller;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class AuthControllerTest {
-
-}
