+++ src/test/java/backend/sumnail/medium/AuthControllerTest.java
@@ -1,0 +1,5 @@
+package backend.sumnail.medium;
+
+class AuthControllerTest {
+
+}
